Swim sounds added
I created a very basic animator for Kame and synced it with Wwise through event callers in the animation frames. The framework now exists for it to with ease be synced with new and "real" animations in the future.
Other than what is updated here, I learned alot of new things within Wwise too