Changed in some prefabs concerning the fishes and the scripts concerning corals and fish
More specifically did I fix so that you can choose which fish colour to deposit and you can only deposit the amount of fish the coral wants(not more, but less is okay). Its still a little bit klicking but until My is done with her UI stuff i think this works good.