Finally fixed so that you can leave fish to safezones
I over complicated things as usual. I could just use the same scripts as i did with the coral and just update the UI when the player enters the trigger of the object etc. What I have left to do is renamne the Coral script to something more general and fix so that the player can pick up the fish in the safe zones again.