Fish deposits automatically to corals  now and is safe from enemies
Made a new prefab that is FunctionalSafezone. Basicclly all I did was go into safezone prefab and change the name of FunctionalCoral to FunctionalSafezone and made it a prefab. I also fixed two of my testing scenes so they work correctly and fixed so that fish deposits automatically to corals and that they are safe zones.