Fixed some more smol fixes
I fixed so that when fish respawn they get a particle effect.  Somethign that needs to be done now tho is to inactivate the particle system when the fish has been picked up. I don't have time to do it now but might check on it tomorrow if I have time.